RecipeCollection.java: make it testable with empty constructor

+++ LOOPs/app/src/main/java/com/example/loops/modelCollections/RecipeCollection.java
@@ -64,7 +64,8 @@ public class RecipeCollection {
         if (!allRecipes.contains(recipe)){
             allRecipes.add(recipe);
         }
-        database.addDocument(recipe);
+        if (database != null)
+            database.addDocument(recipe);
     }
 
     /**
@@ -74,7 +75,8 @@ public class RecipeCollection {
      */
     public boolean deleteRecipe(int indexToDelete){
         try {
-            database.deleteDocument(allRecipes.get(indexToDelete));
+            if (database != null)
+                database.deleteDocument(allRecipes.get(indexToDelete));
             allRecipes.remove(indexToDelete);
         } catch (Exception e) {
             return false;
@@ -107,7 +109,8 @@ public class RecipeCollection {
      */
     public boolean updateRecipe(int recipeInd, Recipe newRecipe){
         try {
-            database.updateDocument(allRecipes.get(recipeInd), newRecipe);
+            if (database != null)
+                database.updateDocument(allRecipes.get(recipeInd), newRecipe);
             allRecipes.set(recipeInd, newRecipe);
         }
         catch (Exception e) {
